+++ amr_bt_navigator/include/amr_bt_navigator/bt_navigator.hpp
@@ -13,6 +13,7 @@
 #include "ament_index_cpp/get_package_share_directory.hpp"
 #include "amr_msgs/action/navigate_to_pose.hpp"
 #include "amr_msgs/msg/local_plan_status.hpp"
+#include "amr_msgs/msg/localization_status.hpp"
 #include "amr_msgs/msg/motion_command.hpp"
 #include "amr_msgs/msg/motion_status.hpp"
 #include "amr_msgs/srv/clear_costmap.hpp"
@@ -56,9 +57,11 @@ private:
   void handle_current_pose(const geometry_msgs::msg::PoseStamped::SharedPtr message);
   void handle_motion_status(const amr_msgs::msg::MotionStatus::SharedPtr message);
   void handle_local_plan_status(const amr_msgs::msg::LocalPlanStatus::SharedPtr message);
+  void handle_localization_status(const amr_msgs::msg::LocalizationStatus::SharedPtr message);
   geometry_msgs::msg::PoseStamped get_current_pose_copy() const;
   amr_msgs::msg::MotionStatus get_motion_status_copy() const;
   amr_msgs::msg::LocalPlanStatus get_local_plan_status_copy() const;
+  amr_msgs::msg::LocalizationStatus get_localization_status_copy() const;
   bool is_navigator_ready(std::string & error_message) const;
   bool wait_for_planner_service(std::string & error_message);
   bool wait_for_recovery_services(std::string & error_message);
@@ -91,10 +94,12 @@ private:
   rclcpp::Subscription<geometry_msgs::msg::PoseStamped>::SharedPtr current_pose_subscription_;
   rclcpp::Subscription<amr_msgs::msg::MotionStatus>::SharedPtr motion_status_subscription_;
   rclcpp::Subscription<amr_msgs::msg::LocalPlanStatus>::SharedPtr local_plan_status_subscription_;
+  rclcpp::Subscription<amr_msgs::msg::LocalizationStatus>::SharedPtr localization_status_subscription_;
   rclcpp_lifecycle::LifecyclePublisher<amr_msgs::msg::MotionCommand>::SharedPtr motion_command_publisher_;
   std::string navigate_action_name_;
   std::string command_topic_;
   std::string current_pose_topic_;
+  std::string localization_status_topic_;
   std::string motion_status_topic_;
   std::string local_plan_status_topic_;
   std::string plan_recovery_service_;
@@ -110,9 +115,11 @@ private:
   geometry_msgs::msg::PoseStamped current_pose_;
   amr_msgs::msg::MotionStatus latest_motion_status_;
   amr_msgs::msg::LocalPlanStatus latest_local_plan_status_;
+  amr_msgs::msg::LocalizationStatus latest_localization_status_;
   bool has_current_pose_;
   bool has_motion_status_;
   bool has_local_plan_status_;
+  bool has_localization_status_;
   mutable std::mutex navigator_mutex_;
   mutable std::mutex active_goal_mutex_;
   std::weak_ptr<GoalHandleNavigateToPose> active_goal_handle_;